fix bug

+++ app/src/main/java/com/kominfo/anaksehat/models/GiveBirth.java
@@ -15,7 +15,7 @@ public class GiveBirth {
     private int mother_condition_id;
     private String mother_condition_name;
     private String remarks;
-    private int child_id;
+    private long child_id;
     private int user_id;
 
     public long getId() {
@@ -122,11 +122,11 @@ public class GiveBirth {
         this.user_id = user_id;
     }
 
-    public int getChild_id() {
+    public long getChild_id() {
         return child_id;
     }
 
-    public void setChild_id(int child_id) {
+    public void setChild_id(long child_id) {
         this.child_id = child_id;
     }
 }
